% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distDB.R
\name{distDB}
\alias{distDB}
\title{# Davis-Bouldin Type Distance}
\usage{
distDB(mean, Sigma)
}
\arguments{
\item{mean}{A matrix of mean values.}

\item{Sigma}{A list of variance-covariance matrix.}
}
\value{
A distance matrix.
}
\description{
This function calculates Davis-Bouldin type distance.
}
