% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distDB2.R
\name{distDB2}
\alias{distDB2}
\title{# Davis-Bouldin Type Distance}
\usage{
distDB2(X, hmmvb, clust)
}
\arguments{
\item{X}{A data matrix.}

\item{hmmvb}{An object of class 'HMMVB'}

\item{clust}{An object of class 'HMMVBclust'}
}
\value{
A distance matrix.
}
\description{
This function calculates Davis-Bouldin type distance from the estimated GMM components of the output of HMM-VB.
}
