% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitHmmvb.R
\name{fitHmmvb}
\alias{fitHmmvb}
\title{Hidden Markov Model on Variable Blocks (HMM-VB).}
\usage{
fitHmmvb(X, C, vbs, numst = rep(0, length(vbs$vb)))
}
\arguments{
\item{X}{A data matrix.}

\item{C}{The number of clusters.}

\item{vbs}{Variable block structure. The output of constVB.}
}
\value{
A list of estimated parameters for HMM-VB.
}
\description{
This function fitt the model HMM-VB on a given data set.
}
\examples{
# Data generation
set.seed(1)
dat = genData2(n=300,p1=100,p2=100,C=5,rep=1) 
X = dat[[1]]$X_total
Y = dat[[1]]$z
n = dim(X)[1];p = dim(X)[2]
# The number of clusters.
C = 5

# Maximum block size is set 5\% of the total dimension.
max.vb.size = 10
# Calculate Mutual Information.
pwmi = pairwiseMI(X)
# Variable block construction by ESS-DFS.
vbs = constVB(X,pwmi,max.vb.size)

# Fitting HMM-VB with 5 components.
fit = fitHmmvb(X,C,vbs)

# Semi-clusters
semi.cls = semicls(X,fit)
# Variable block selection by a bimodality test.
chosen.vb = semi.cls$chosen.vb
# Reduce the model structure
red.dat = reduceVB(X,fit,chosen.vb)
red.X = red.dat$X
red.vbs = red.dat$vbs

# Re-estimation of the HMM-VB model with reduced dimensions
re.fit = fitHmmvb(red.X,C,red.vbs)
# Final clustering
final.cls = finalcls(red.dat$X,re.fit,C,5)
}
