% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genData.R
\name{genData}
\alias{genData}
\title{Generating multi-dimensional Gaussian random partition.}
\usage{
genData(
  n,
  p1,
  p2,
  C,
  relev.var = c(10, 50),
  irrel.var = c(100, 400),
  overlap = FALSE,
  BClevel = 10^-5,
  MinSize = 1/(5 * C),
  rep = 1
)
}
\arguments{
\item{n}{Size of observations.}

\item{p1}{Dimension of relevant variables.}

\item{p2}{Dimension of irrelevant variables.}

\item{C}{The number of clusters.}

\item{relev.var}{Range of positive number. Default is [10,50]. Range for variances of relevant variables.}

\item{irrel.var}{Range of positive number. Default is [100,400]. Range for variances of irrelevant variables.}

\item{overlap}{Logical. If it is True, the algorithm allows clusters on relevant variable space to overlap each other. Overlap is evaluated by Bhattacharyya coefficient.}

\item{BClevel}{Threshold of Bhattacharyya coefficient. Default is 10^-5.}

\item{MinSize}{Minimum proportion allowed for each cluster. Default is 1/(5*C).}
}
\value{
A synthesized data matrix.
}
\description{
This function generates multi-dimensional Gaussian random partition with noisy variables for simulation study.
}
\examples{
sim1 = genData(n=300,p1=100,p2=100,C=5,rep=1)
sim2 = genData2(n=300,p1=100,p2=100,C=5,rep=1) 
}
