% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genData2.R
\name{genData2}
\alias{genData2}
\title{Generating multi-dimensional Gaussian partition based on given parameters.}
\usage{
genData2(n, p1, p2, C, rep = 10)
}
\arguments{
\item{n}{Size of observations.}

\item{p1}{Dimension of relevant variables.}

\item{p2}{Dimension of irrelevant variables.}

\item{C}{The number of clusters.}

\item{rep}{The number of data sets.}
}
\value{
A list of the synthesized data matrix.
}
\description{
This function generates multi-dimensional Gaussian partition with noisy variables for simulation study based on pre-specified mean and covariance matrix parameters.
}
\examples{
sim1 = genData(n=300,p1=100,p2=100,C=5,rep=1) 
sim2 = genData2(n=300,p1=100,p2=100,C=5,rep=1) 
}
