% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genData3.R
\name{genData3}
\alias{genData3}
\title{Generating multi-dimensional Gaussian random partition.}
\usage{
genData3(
  n,
  relev.blocks,
  irrel.blocks,
  C,
  relev.var = c(100, 200),
  irrel.var = c(100, 200),
  df = 4,
  overlap = FALSE,
  BClevel = 10^-5,
  MinSize = 1/(5 * C),
  rep = 1
)
}
\arguments{
\item{n}{Size of observations.}

\item{relev.blocks}{A vector of positive integers indicating the size of each relevant block.}

\item{irrel.blocks}{A vector of positive integers indicating the size of each irrelevant block.}

\item{C}{The number of clusters.}

\item{relev.var}{Range of positive number. Default is [10,50]. Range for variances of relevant variables.}

\item{irrel.var}{Range of positive number. Default is [100,400]. Range for variances of irrelevant variables.}

\item{overlap}{Logical. If it is True, the algorithm allows clusters on relevant variable space to overlap each other. Overlap is evaluated by Bhattacharyya coefficient.}

\item{BClevel}{Threshold of Bhattacharyya coefficient. Default is 10^-5.}

\item{MinSize}{Minimum proportion allowed for each cluster. Default is 1/(5*C).}
}
\value{
A synthesized data matrix.
}
\description{
This function generates multi-dimensional Gaussian random partition with noisy variables for simulation study.
}
\examples{
sim1 = genData(n=300,p1=100,p2=100,C=5,rep=1)
sim2 = genData2(n=300,p1=100,p2=100,C=5,rep=1) 
sim3 = genData3(n=1000,relev.blocks=c(200,200,200,200,200),irrel.blocks=1000,relev.var=c(50,50),irrel.var=c(50,50),C=5,rep=1) 
}
