% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDclustVS-package.R
\docType{package}
\name{HDclustVS-package}
\alias{HDclustVS-package}
\alias{_PACKAGE}
\alias{HDclustVS}
\title{Block-wise Variable Selection for Clustering via Latent States of Mixture Models}
\description{
\pkg{HDclustVS} is an R package for new block-wise variable selection methods for clustering, HMM-VB-VS and GMM-VB-VS, which exploit the latent states of the hidden Markov model on variable blocks or Gaussian mixture model. The variable blocks are formed by early-stop-and-sorted-depth-first-search (ESS-DFS) on a dendrogram created based on the mutual information between any pair of variables. Then, the variable selection is conducted by an independence test between the latent states and semi-clusters which are the smaller clusters that will be further grouped into final clusters.
}
\author{
\strong{Maintainer}: Beomseok Seo \email{bzs32@psu.edu}

Authors:
\itemize{
  \item Lin Lin \email{llin@psu.edu}
  \item Jia Li \email{jiali@psu.edu}
}

}
\keyword{internal}
