% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LPS}
\alias{LPS}
\title{A single-cell RNA-sequencing data from primary mamalian cells}
\format{
A list containing four components
  \item{cell}{cell names}
  \item{cell.stage}{true labels of cells}
  \item{gnames}{gene names}
  \item{expr}{expression data}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/19729616}
}
\usage{
LPS
}
\description{
A dataset consisting of 131 primary mammalian cells with 534 genes. This data set contains only two cell clusters: unstimulated and stimulated cells.
}
\keyword{datasets}
