% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MEF}
\alias{MEF}
\title{A single-cell RNA-sequencing data from mouse cells.}
\format{
A list containing four components
  \item{cell}{cell names}
  \item{cell.stage}{true labels of cells}
  \item{gnames}{gene names}
  \item{expr}{expression data}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/19729616}
}
\usage{
MEF
}
\description{
A dataset consisting of 92 single cells with 35 genes from two different mouse cell types: embryonic stem cells (ES R1) and embryonic fibroblasts (MEFs).
}
\keyword{datasets}
