% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeSize.R
\name{nodeSize}
\alias{nodeSize}
\title{Computing the size of the subtree of a node.}
\usage{
nodeSize(M)
}
\arguments{
\item{M}{A matrix of merges of nodes, which is obtained from hclust() output.}
}
\value{
A vector with the size of leaves-set of each node.
}
\description{
This function calculates the size of leaves-set of a node.
}
