% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseMI.R
\name{pairwiseMI}
\alias{pairwiseMI}
\title{Pairwise Mutual Information.}
\usage{
pairwiseMI(X, numBins1 = 10, numBins2 = 10)
}
\arguments{
\item{X}{A data matrix.}
}
\value{
A matrix of pairwise mutual information.
}
\description{
This function calculates pairwise mutual information between variables.
}
\examples{
# Data generation
set.seed(1)
dat = genData2(n=300,p1=100,p2=100,C=5,rep=1) 
X = dat[[1]]$X_total
Y = dat[[1]]$z
n = dim(X)[1];p = dim(X)[2]
# The number of clusters.
C = 5
# Maximum block size is set 5\% of the total dimension.
max.vb.size = 10
# Calculate Mutual Information.
pwmi = pairwiseMI(X)
}
