% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCls.R
\name{plotCls}
\alias{plotCls}
\title{Visulization of a partition on 2 dimensional space with convex hull.}
\usage{
plotCls(
  X,
  z,
  title = "",
  xlab = "",
  ylab = "",
  legend.title = "",
  no.legend = F
)
}
\arguments{
\item{X}{A matrix of data.}

\item{z}{Cluster labels of data.}
}
\value{
None
}
\description{
This function plots a partition on 2 dimensional principal components space with convex hull.
}
\examples{
# Data generation
set.seed(1)
dat = genData2(n=300,p1=100,p2=100,C=5,rep=1) 
X = dat[[1]]$X_total
Y = dat[[1]]$z
n = dim(X)[1];p = dim(X)[2]

PCA = prcomp(dat[[1]]$X_relev)
Xr = prcomp(dat[[1]]$X_relev)$x[,1:2]
plotCls(Xr,Y,title="Relevant var.")
Xi = prcomp(dat[[1]]$X_irrel)$x[,1:2]
plotCls(Xi,Y,title="Irrelevant var.")
Xt = prcomp(dat[[1]]$X_total)$x[,1:2]
plotCls(Xt,Y,title="Total var.")
}
