% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmvt.R
\name{rtmvt}
\alias{rtmvt}
\title{Generating multi-dimensional Gaussian random partition.}
\usage{
rtmvt(
  n,
  mean,
  sigma,
  df = 3,
  type = "shifted",
  lower = rep(-inf, length(mean)),
  upper = rep(inf, length(mean))
)
}
\arguments{
\item{n}{Size of observations.}

\item{mean}{Mean vectors.}

\item{sigma}{Covariance matrix.}

\item{df}{Degree of freedom.}

\item{type}{Type of the noncentral multivariate t distribution. Refer to mvtnorm::rmvt().}

\item{lower}{Lower bound of truncated t-distribution.}

\item{upper}{Upper bound of truncated t-distribution.}
}
\value{
truncated t-distribution samples
}
\description{
This function generates multi-dimensional Gaussian random partition with noisy variables for simulation study.
}
