% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchTreeCls.R
\name{search.cls}
\alias{search.cls}
\title{Clustering Latent State Configurations by BFS Search.}
\usage{
search.cls(tree, dist, num.cls = NULL, min.size = 5)
}
\arguments{
\item{tree}{An object of hclust result.}

\item{dist}{The distance measures between nodes for the tree.}

\item{num.cls}{The number of cluster. If it is NULL, it finds semi-cluster. If it is specified as an integer, it finds clusters with the number of groups.}

\item{min.size}{The minimum size of each cluster. Default is 5.}
}
\value{
A vector of cluster indices for each leaf node.
DD = dist(X)
min.size.cls = 5
tree = hclust(DD)
cls.ix = search.cls(tree,DD,C,min.size.cls)
}
\description{
This function searches clusters by BFS search from a dendrogram tree.
}
