% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim1}
\alias{sim1}
\title{Simulated toy data}
\format{
A list containing four components
  \item{X_total}{total data}
  \item{X_relev}{relevant data}
  \item{X_irrel}{irrelevant data}
  \item{z}{true labels of clusters}
}
\usage{
sim1
}
\description{
A dataset containing 300 samples and 200 variables.
}
\keyword{datasets}
