% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viterbi.R
\name{viterbi}
\alias{viterbi}
\alias{TransProb}
\alias{EmissProb}
\alias{Viterbi}
\alias{GaussParam}
\title{Cross Validation for Variable Selection Threshold}
\usage{
TransProb(X, fit)

EmissProb(X, fit)

Viterbi(TP, EP, NN, TT, KK)

GaussParam(HC, VV, BD, NN, TT, KK)
}
\arguments{
\item{X}{A data matrix}

\item{fit}{HMM-VB or GMM-VB output}

\item{TP}{Transition probability list. The output of TransProb().}

\item{EP}{Emission probability list. The output of EmissProb().}

\item{NN}{Sample size. dim(X)[1].}

\item{TT}{The length of variable blocks.}

\item{KK}{A vector of the number of components for variable blocks.}

\item{HC}{Hidden Markov chain}

\item{VV}{Viterbi path. The output of Viterbi().}

\item{BD}{A vector of the number of variables for each variable block.}
}
\value{
TransProb gives a list of transition probability.

EmissProb gives a list of emission probability.

Viterbi gives a matrix of viterbi paths.

GaussParm gives a list of estimated mixture model parameters for each sample.
\item{WW}{Mean values of each sample corresponding to MAP state configuration.}
\item{SS}{Variance values of each sample corresponding to MAP state configuration.}
}
\description{
This function finds a threshold for variable selection using cross validation.
}
