% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{YAN}
\alias{YAN}
\title{A single-cell RNA-sequencing data from human preimplantation embryos and human embryonic stem cells.}
\format{
A list containing four components
  \item{cell}{cell names}
  \item{cell.stage}{true labels of cells}
  \item{gnames}{gene names}
  \item{expr}{expression data}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/23934149}
}
\usage{
YAN
}
\description{
A dataset consisting of 124 individual cells with 3,840 genes from human preimplantation embryos and human embryonic stem cells. There are 7 cell clusters: Oocyte and Zygote, 2-cell, 4-cell, 8-cell, Morula, Late blastocyst, hESC.
}
\keyword{datasets}
